refactor(app): move route-local containers out of _components
src/app/AGENTS.md places a route-local client component in _containers/
as soon as it owns SWR fetching, a Server Action binding, or a store
derived side effect. Thirteen components had drifted into _components/
while doing exactly that, so the folder no longer told a reader whether
a file was a view or a data owner.

Move them and their co-located component tests, and rewrite the affected
import paths. No behaviour changes: renames only, plus specifier edits.

AdminReviewsTemplate becomes AdminReviewsTable in the same pass. The
Template suffix is reserved by src/ui/components/templates/AGENTS.md for
pure page bodies that take finished content as props, and this one owns
deleteReviewByAdmin, toast and router.refresh — so the name asserted a
purity the file never had. Sibling containers carry no suffix, and the
admin section already speaks in ProductTableRow terms.

(admin)/admin/reviews/_components/ is now empty and removed.
