+++ src/app/(admin)/admin/reviews/page.tsx
@@ -5,7 +5,7 @@ import { getAdminReviewsPageService } from "@/services/review";
 import { adminReviewListRequestSchema } from "@/core/schemas/request/adminReviewList.schema";
 import { decodeCursor } from "@/core/utils/cursor";
 import { validateAndFlatten } from "@/core/utils/validate-and-flatten";
-import { AdminReviewsTemplate } from "@/app/(admin)/admin/reviews/_components/AdminReviewsTemplate";
+import { AdminReviewsTable } from "@/app/(admin)/admin/reviews/_containers/AdminReviewsTable";
 
 // 커서는 URL이 소유하므로 어떤 입력이 와도 throw하지 않는다 — 형식이 깨진 cursor는
 // decodeCursor가 걸러 "커서 없음"으로 떨어뜨린다.
@@ -33,7 +33,7 @@ const ReviewsPage = async ({
   const { cursor } = resolveFilters(await searchParams);
   const page = await getAdminReviewsPageService({ cursor });
 
-  return <AdminReviewsTemplate page={page} cursor={cursor} />;
+  return <AdminReviewsTable page={page} cursor={cursor} />;
 };
 
 export default ReviewsPage;
