+++ src/app/(main)/(products)/products/[category]/[id]/_containers/ProductLikeBadge.component.test.tsx
