+++ src/app/(main)/(products)/products/[category]/[id]/_containers/ProductViewTracker.tsx
