+++ src/app/(preview)/preview/[publicKey]/_components/GuestbookSection.tsx
@@ -1,6 +1,6 @@
 import { routes } from "@/core/domain/routes";
 import { DemoGuestbookSection } from "./DemoGuestbookSection";
-import { LiveGuestbookSection } from "./LiveGuestbookSection";
+import { LiveGuestbookSection } from "../_containers/LiveGuestbookSection";
 
 export function GuestbookSection({ publicKey }: { publicKey: string }) {
   if (publicKey === routes.preview.samplePublicKey) {
